+++ src/locales/index.ts
@@ -3,11 +3,13 @@ import { addLocaleData } from 'react-intl';
 import * as frLocaleData from 'react-intl/locale-data/fr';
 import * as koLocaleData from 'react-intl/locale-data/ko';
 
+// Add locale data
 addLocaleData([
   ...frLocaleData,
   ...koLocaleData,
 ]);
 
+// Add translation data
 const translations = {
   en: {},
   fr: require('./lang/fr.json'),