+++ src/plugins/core/widgets/time/Time.tsx
@@ -1,3 +1,4 @@
+import { DateTime } from 'luxon';
 import * as React from 'react';
 import Analogue from './Analogue';
 import Digital from './Digital';
@@ -5,10 +6,11 @@ import Digital from './Digital';
 interface Props {
   hour12: boolean;
   mode: string;
+  timezone?: string;
 }
 
 interface State {
-  time: Date;
+  time: DateTime;
 }
 
 class Time extends React.PureComponent<Props, State> {
@@ -16,17 +18,19 @@ class Time extends React.PureComponent<Props, State> {
     mode: 'digital',
     hour12: false,
   };
-  state: State = {
-    time: new Date(),
-  };
+  state: State = { time: this.getDateTime() };
   private interval: number;
 
   componentWillMount() {
-    this.interval = window.setInterval(() => this.setState({ time: new Date() }), 1000);
+    this.interval = window.setInterval(this.tick, 1000);
   }
 
   componentWillUnmount() {
     window.clearInterval(this.interval);
+  }
+
+  componentWillReceiveProps(props: Props) {
+    this.tick(props);
   }
 
   render() {
@@ -35,6 +39,16 @@ class Time extends React.PureComponent<Props, State> {
     }
 
     return <Digital time={this.state.time} hour12={this.props.hour12} />;
+  }
+
+  private tick = (props: Props = this.props) => {
+    this.setState({ time: this.getDateTime(props) });
+  }
+
+  private getDateTime({ timezone }: Props = this.props) {
+    return timezone
+      ? DateTime.local().setZone(timezone)
+      : DateTime.local();
   }
 }