+++ src/plugins/core/widgets/time/TimeSettings.tsx
@@ -1,9 +1,12 @@
+import { Info } from 'luxon';
 import * as React from 'react';
 import { Settings } from '../../../interfaces';
+const timezones: string[] = require('./timezones.json');
 
 interface Props {
   mode: string;
   hour12: boolean;
+  timezone?: string;
   onChange: (settings: Settings) => void;
 }
 
@@ -51,6 +54,22 @@ class TimeSettings extends React.PureComponent<Props> {
           {' '}
           24-hour digital
         </label>
+
+        {(Info.features() as any).zones && ( // PR waiting to fix this type
+          <label>
+            Timezone
+
+            <select
+              value={this.props.timezone}
+              onChange={event => this.props.onChange({ timezone: event.target.value})}
+            >
+              <option value={''}>Browser default</option>
+              {timezones.map(timezone =>
+                <option key={timezone} value={timezone}>{timezone.replace('_', ' ')}</option>
+              )}
+            </select>
+          </label>
+        )}
       </div>
     );
   }